requirements.txt
configparser==5.3.0
discord.py==2.1.0
SQLAlchemy==1.4.42
PyMySQL==1.0.2
cffi==1.15.1
Flask==2.2.2
Flask-SSE==1.0.0
gevent==22.10.2
requests==2.28.1
APScheduler==3.9.1.post1
cryptography==40.0.2
quart==0.18.4
quart_cors==0.6.0
hypercorn==0.14.4
pytimeparse==1.1.8
PyYAML==6.0.1